Rename to AddReactionsPrompt
