+++ src/blocks/index.js
@@ -1,2 +1,2 @@
-export { default as ReactionPrompt } from './reaction_prompt'
+export { default as AddReactionsPrompt } from './add_reactions_prompt'
 export { default as SubmissionLayout } from './submission_layout'
